10.8.2 fix(lua): Use container include properly:
- `src/lua/pandoc-argmap.lua`: Use pandoc.pipe with pandoc command
itself but no input, and the container include file as the template,
so that template variables are populated with meta data from the
config files etc.

Use container include file properly: use pandoc.pipe with pandoc command itself but no input, and the container include file as the template, so that template variables are populated with meta data from the config files etc.